in the dream,█was something. in the dream someone asked if they could listen to me and█wasn't the one to ask. in the dream there were people waiting to hear what█had to say and█was running late for something█wasn't actually scared of. in the dream there was a dog, in a cage, made flimsy, and sharp, and all█did was stand and watch, and feed him a bumblebee. in the dream█was something in the dream there were people waiting in the dream there was a dog and there was a bee. in the dream█was in egypt at a crossroads and speaking with a man about the fight against ephemera in the dream█was as solid as the ancient stone█stared at and in the dream█was something in the dream someone asked if they could listentomeandiwasnttheonetoaskinthedreamtherewassomebodywaitingthereweremanypeopletheywerewaitingtohearwhatihadtosayandididnthavetobegtoexistinthedreamiwassomething.gnihtemossawimaerdehtnitsixeotgebotevahtndididnayasotdahitahwraehotgnitiawerewyehtelpoepynamerewerehtgnitiawydobemossawerehtmaerdehtniksaotenoehttnsawidnaemotnetsil listen to me█am telling you that█love you. adrian are you listening to me█know you have no ears but who does anyway. in the dream someone asked if they could listen to me and█wasn't the one to ask. in the dream there was a lonely giant dog and█fed him a dead bumble bee. in the dream█wasn't wishing for a time with words because█was the time the words were written in and in the dream█was something and people were waiting for me.█am waiting for me to catch up to you and█am waiting for a time with words in the dream█hadn't a thing except track of the time before█had to speak to crowds of people waiting for me patiently. listen.█am listening for all the many ways█loved me in the dream█had a home a husband and nothing felt like strangling there. in the dream there was a stream and it was beautiful.█was something in the dream█fed a dog a bumblebee and█don't need to know what this means. because█am still dreaming but here i'm nothing and ask for doctors landlords rapists to hear me when█say to help me in or please get out in the dream█was something in the dream someone asked if they could listen and█didn't have to ask in the dream there were people waiting for me they were waiting and god callie what are you waiting for except someone to hear you scream in the dream there was a dog and█was the dog and the cage and the bee that was eaten and in the dream█was something.█was something, in the dream
in the wake█am something.█am awake and█am something. in the wake of everything█am still something.█am someone in the wake of silencing.█am speaking and█am the wake of the silence.█am speaking and█will awake to the spoken.█am speaking and█do not care if you listen.█am speaking and█will keep speaking as spoken to.█am still speaking and█do not care who it's speaking to.█am the speaking and wake of the spoken to.█am the speaker at wake of the spokes of you.█will keep speaking and speak out the name of you.█have my name and it speaks as a stream that is beautiful.█am a name and that name it is waking you.█am the name of a dream to awake from and spoken new.█speak my name and it dreams me as streaming through. and█am awake to the name-taking plain of you.█swallow your name to spit lightning. i’m striking you. i, as ephemera, a weight as cold stone of the tomb.█am the stream. and the bee. and the dog. that dog that█fed, it is me. and the bee that█fed it to.█am the dog's drool as its eyes stay transfixed on the hand that may offer its next chance at pitiful food.█am the dead dog and bee sting on loud tongues; defiantly dying in place of just lying and taking a bite from the dog's chipped and cowardly tooth. and that█once fed it.█am not ashamed of the waste of myself to abuse.█am the poisonous spear of the hearing you think of as food. and█am that█fed it, too.█am my name and the sound of my dreams as they broke in two.█AM the WAVE of the sea and the waves as they BREAK TO YOU.█am a goodBYE and the NAME of the SACRED LIFE TAKEN from me far too soon.█AM THE SPEAKING OF SILENCE THAT WAKES UP A FEAR IN THE SMALL PARTS INSIDE OF YOU.█AM THE UNJUSTLY UNSPOKEN, THE DRY BROKEN BONE, THE TORN AND WORN RAPED FLESH AND WET SIN OF SINEW.█AM DISCOMFORT. POST-BLOWJOB BRUISED KNEES TO PAY RENT FOR A SQUAT ROTTED WITH FILTH AND DISEASE. AND█AM THE SWARM AS IT GATHERS ROUND SHIT THAT YOU BREATHE.█AM WHAT'S LEFT OF THE FALLEN, BEEN TAKEN, AND█AM THE FALLING AND RAKING OF LEAVES.█AM THAT█AM GIVING BIRTH TO YOUR SHAME.█AM BIRTHED AND REMADE BY YOUR HATE.█AM SHAMED,█AM SHAME AND THAT SHAMEFUL AND DEEPEST KEPT SECRET DARK SIDE OF YOU.█AM THE SNAPPING OF BRANCHES WHO ONLY GREW WIDE TO BE USED AS SUPPORT FOR A NOOSE.█AM THE UNBREATHABLE PROOF. THE MIRROR OF YOU, WHAT YOU DID, AND WILL NEVER DO.█AM THE FUCKING GREAT FUMES OF THE *TRUTH.*█WILL NOT BE LESS GRUESOME, MORE MEEK, FOR THE SAKE OF THE INTERNAL STATE OF THE CHURNING BILE WASTE THAT'S INSIDE OF YOU.█AM LITURGY SUTURED AND STAPLED SCALP SURGICAL SCAR TISSUE. AN ABSOLUTE ZERO THIRD UNFINISHED DEGREE UNEASE AND█AM NOT MADE TO PLEASE YOU.█LICK AIR GLOSSOLALIC AS FIRE WITHOUT THOUGHT OF THE HIDEOUS STRUCTURES CONSUMED.█AM THE WORD AND█AM ITS LETTER TO THE GODDESS YOU TRIED TO MAKE ME FORGET. BUT█REMEMBER. AND█WILL NEVER FORGET HER.█AM UNEXORCISABLE. DESPISED AND DESPISING,█WAS BUILT TO BE BROKEN, PROTECTION, AS WARD, AS A WARNING. OF WARDEN, AS FALSE START OF LOSS, AS A WAR.█WAS BORN AS AN ART FORM BEGINNING AS NEW. A GARDEN OF HANGING SWORDS PROMISING JUSTICE IN ENDINGS AND IMPENDING DOOM. AS A FUSE.█AM MUSIC. AND FORCE.█AM LOUDNESS! WILL NO LONGER STAND FOR ENFORCED QUIETUDE!█AM STANDING UP NOW! SO JUST SIT DOWN OR GIVE YOUR TIRED MOUTH SOMETHING BETTER TO DO!█AM HER! AND HERE SPEAKING. HERE. SPEAKING! SO HEAR ME! OR NOT! WHATEVER YOU CHOOSE! YOU CAN SIT DOWN UNCOMFORTABLY!█REFUSE TO MAKE SPACE IN THIS PRISON THAT YOU CALL A ROOM!█AM HERE AND I'M STAYING FOR ME AND FOR THEM AND FOR MY RIGHT TO BE STILL AND NOT KILLED SO JUST FUCKING MOVE! YES.█MEAN YOU.█KNOW WHO YOU ARE. BUT DO YOU?█WILL NEVER AGAIN GIVE MY WAY TO A SINGLE SOUL BENT UPON BLOCKING MY VIEW!

SIT!
	DOWN!
			THE
				SHOW'S
						OVER!

██
AM.
	STARTING.

	TO
	SPEAK.
	TO.
	YOU.